Finished Rental Agreement implementation and print function

+++ ToolStore/src/main/java/toolStore/Tool.java
@@ -61,7 +61,7 @@ public class Tool {
         return weekendCharge;
     }
 
-    public double getDailyCharge() {
+    public float getDailyCharge() {
         return dailyCharge;
     }
